Separate installation into two distinct sections
The existing README is not particularly well suited to people who
actually want to *use* a script repository rather than create their own
scripts via their own GitHub fork.  This commit improves that.

An HTTPS clone URL is also used for the user installation section.  This
URL is actually usable in a systems administration context where the
installation environment is not registered with GitHub.