Move JVM memory settings section to the server installation page
As we expect all the sysadmins to read this particular section, the JVM memory
settings section is now directly included into the Server Installation
sections for UNIX and Windows platforms. The link to the advanced server
configuration section is preserved for informative purposes only alongside
with the link to the forum thread.