Allow passing extra channel rendering options to Make_Movie script (see #12037)
Reusing the Channel keyword argument, now accepting a list of strings instead of a list of integers. Backward compatible with the previous format of channel ids (1 based) to turn on exclusively.
ex: Channels=2,3

Extended the accepted format to allow for passing channel window range and color, with the format 'idx|min:max' having both '|min:max' and '' optional, but when both are present order must be respected.
ex: Channels=2|0:250,3|50:100