Move jvm settings to adv. configuration (See #12311)
Now that the default memory configuration need not
be modified by default, it's less critical that the
info be embedded in the main installation page. Also,
this text will need to grow substantially in order to
make the usage clear, and repeating that between unix
and windows seems error prone.

Conflicts:
	omero/sysadmins/server-advanced-configuration.txt