Add top-level Git mailmap to normalize commit author variants
This mapping file aims to associate all commits from the same contributor
to a single "First Name Last Name <email>" author line. The
University of Dundee or the Glencoe Software email address are used
as the canonical address for current employees of these institutions.
For contributors with an executed CLA, the email address specified
in the  CLA is used as the canonical one.